A bit more refactoring of packages
