+++ src/api/java/com/minecolonies/api/research/factories/IResearchEffectFactory.java
@@ -3,7 +3,7 @@ package com.minecolonies.api.research.factories;
 import com.minecolonies.api.colony.requestsystem.factory.FactoryVoidInput;
 import com.minecolonies.api.colony.requestsystem.factory.IFactory;
 import com.minecolonies.api.colony.requestsystem.factory.IFactoryController;
-import com.minecolonies.api.research.IResearchEffect;
+import com.minecolonies.api.research.effects.IResearchEffect;
 import org.jetbrains.annotations.NotNull;
 
 import static com.minecolonies.api.util.constant.Constants.PARAMS_RESEARCH_EFFECT;
