+++ src/api/java/com/minecolonies/api/research/util/ResearchState.java
@@ -1,7 +1,7 @@
 package com.minecolonies.api.research.util;
 
 /**
- * Class which defines the possible states of a research.s
+ * Class which defines the possible states of a research.
  */
 public enum ResearchState
 {
