+++ src/main/java/com/minecolonies/coremod/research/ResearchEffectManager.java
@@ -1,7 +1,7 @@
 package com.minecolonies.coremod.research;
 
 import com.minecolonies.api.research.effects.IResearchEffectManager;
-import com.minecolonies.api.research.IResearchEffect;
+import com.minecolonies.api.research.effects.IResearchEffect;
 import com.minecolonies.api.util.Log;
 import org.jetbrains.annotations.NotNull;
